removed u_int, explicitly casted

+++ game/src/Renderer.cpp
@@ -3,9 +3,9 @@
 void Renderer::RenderText( const Text& text )
 {
     DrawText( text.text.c_str(),
-              text.position.x,
-              text.position.y,
-              text.size,
+        static_cast<int>(text.position.x),
+        static_cast<int>(text.position.y),
+        static_cast<int>(text.size),
               text.color );
 }
 
@@ -17,5 +17,5 @@ void Renderer::RenderRect( const Rect& rect )
 void Renderer::RenderButton( const Button& button )
 {
     DrawRectangleV( button.position, button.size, button.color );
-    DrawText( button.text.c_str(), button.textPosition.x, button.textPosition.y, button.textSize, button.textColor );
+    DrawText( button.text.c_str(), static_cast<int>(button.textPosition.x), static_cast<int>(button.textPosition.y), button.textSize, button.textColor);
 }
